Fixed -Wimplicit-int-float-conversion warning.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@92639 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/corelib/ncbitime.cpp
@@ -2315,7 +2315,7 @@ CTimeSpan CTime::DiffTimeSpan(const CTime& t) const
 
 void CTimeSpan::Set(double seconds)
 {
-    if (seconds < kMin_Long  || seconds > kMax_Long) {
+    if (seconds < (double)kMin_Long  || seconds > (double)kMax_Long) {
         NCBI_THROW(CTimeException, eConvert, 
                   "Value " + NStr::DoubleToString(seconds) +
                   " is too big to convert to CTimeSpan");
